PlantillaHibrida/requirementsClustering.txt
click==8.3.2
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
gensim==4.4.0
imbalanced-learn==0.14.1
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.8
nltk==3.9.4
numpy==1.26.4
packaging==26.1
pandas==3.0.2
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
regex==2026.4.4
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sklearn-compat==0.1.5
smart_open==7.6.0
threadpoolctl==3.6.0
tqdm==4.67.3
wrapt==2.1.2
